STEP_2_Authentication_Role_Based_Backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-Cors
Werkzeug
gunicorn
flask-mail
pymysql
google-generativeai
requests
PyPDF2

